* fix bug where adding a new set didn't update the exercise timestamp * merge exerciseSetsProvider into exercisesProvider
