+++ App.tsx
@@ -8,7 +8,6 @@ import functions from "@react-native-firebase/functions";
 import { DefaultTheme, Provider as PaperProvider } from "react-native-paper";
 import { StatusBar } from "expo-status-bar";
 
-import ExerciseSetsProvider from "./data/exerciseSetsProvider";
 import ExercisesProvider from "./data/exercisesProvider";
 
 import Header from "./components/layout/Header";
@@ -68,67 +67,65 @@ function App() {
   const navigationRef = React.useRef();
 
   return (
-    <ExerciseSetsProvider>
-      <ExercisesProvider>
-        <PaperProvider theme={theme}>
-          <StatusBar style="auto" />
-          <NavigationContainer
-            theme={theme}
-            /* @ts-ignore */
-            ref={navigationRef}
-            onReady={() => {
-              routeNameRef.current =
-                /* @ts-ignore */
-                navigationRef.current.getCurrentRoute().name;
-              analytics().logScreenView({
-                screen_name: routeNameRef.current,
-                screen_class: routeNameRef.current,
-              });
-            }}
-            onStateChange={async () => {
-              const previousRouteName = routeNameRef.current;
-              const currentRouteName =
-                /* @ts-ignore */
-                navigationRef.current.getCurrentRoute().name;
+    <ExercisesProvider>
+      <PaperProvider theme={theme}>
+        <StatusBar style="auto" />
+        <NavigationContainer
+          theme={theme}
+          /* @ts-ignore */
+          ref={navigationRef}
+          onReady={() => {
+            routeNameRef.current =
+              /* @ts-ignore */
+              navigationRef.current.getCurrentRoute().name;
+            analytics().logScreenView({
+              screen_name: routeNameRef.current,
+              screen_class: routeNameRef.current,
+            });
+          }}
+          onStateChange={async () => {
+            const previousRouteName = routeNameRef.current;
+            const currentRouteName =
+              /* @ts-ignore */
+              navigationRef.current.getCurrentRoute().name;
 
-              if (previousRouteName !== currentRouteName) {
-                await analytics().logScreenView({
-                  screen_name: currentRouteName,
-                  screen_class: currentRouteName,
-                });
-              }
-              routeNameRef.current = currentRouteName;
+            if (previousRouteName !== currentRouteName) {
+              await analytics().logScreenView({
+                screen_name: currentRouteName,
+                screen_class: currentRouteName,
+              });
+            }
+            routeNameRef.current = currentRouteName;
+          }}
+        >
+          <Stack.Navigator
+            initialRouteName="Exercises"
+            screenOptions={{
+              header: ({ navigation, route, options, back }) => (
+                <Header
+                  navigation={navigation}
+                  route={route}
+                  options={options}
+                  back={back}
+                />
+              ),
             }}
           >
-            <Stack.Navigator
-              initialRouteName="Exercises"
-              screenOptions={{
-                header: ({ navigation, route, options, back }) => (
-                  <Header
-                    navigation={navigation}
-                    route={route}
-                    options={options}
-                    back={back}
-                  />
-                ),
-              }}
-            >
-              <Stack.Group>
-                <Stack.Screen name="Exercises" component={ExercisesScreen} />
-                <Stack.Screen name="Exercise Log" component={Exercise} />
-              </Stack.Group>
-              <Stack.Group screenOptions={{ presentation: "containedModal" }}>
-                <Stack.Screen name="Add Set" component={AddSet} />
-                <Stack.Screen name="Edit Set" component={EditSet} />
-                <Stack.Screen name="Add Exercise" component={AddExercise} />
-                <Stack.Screen name="Edit Exercise" component={EditExercise} />
-              </Stack.Group>
-            </Stack.Navigator>
-          </NavigationContainer>
-          <Footer />
-        </PaperProvider>
-      </ExercisesProvider>
-    </ExerciseSetsProvider>
+            <Stack.Group>
+              <Stack.Screen name="Exercises" component={ExercisesScreen} />
+              <Stack.Screen name="Exercise Log" component={Exercise} />
+            </Stack.Group>
+            <Stack.Group screenOptions={{ presentation: "containedModal" }}>
+              <Stack.Screen name="Add Set" component={AddSet} />
+              <Stack.Screen name="Edit Set" component={EditSet} />
+              <Stack.Screen name="Add Exercise" component={AddExercise} />
+              <Stack.Screen name="Edit Exercise" component={EditExercise} />
+            </Stack.Group>
+          </Stack.Navigator>
+        </NavigationContainer>
+        <Footer />
+      </PaperProvider>
+    </ExercisesProvider>
   );
 }
